tensorflow-federated-variational-autoencoder/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tensorflow-federated-variational-autoencoder"
version = "1.0.0"
description = ""
license = "Apache-2.0"
dependencies = [
    "flwr[simulation]>=1.17.0",
    "flwr-datasets[vision]>=0.5.0",
    "tensorflow>=2.11.1,<2.18.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.flwr.app]
publisher = "allenlin316"

[tool.flwr.app.components]
serverapp = "tensorflow_federated_variational_autoencoder.server_app:app"
clientapp = "tensorflow_federated_variational_autoencoder.client_app:app"

[tool.flwr.app.config]
use-wandb = false
num-server-rounds = 3
local-epochs = 3
learning-rate = 0.001
fraction-fit = 0.25
fraction-evaluate = 0.5
server-device = "cuda:0"
dataset-path = "/workspace/datasets/cic-aa.normal.tls/*.json" # to train and validate the network

[tool.flwr.federations]
default = "local-simulation"

[tool.flwr.federations.local-simulation]
options.num-supernodes = 2
options.backend.init_args.num_gpus = 1 # Expose a single GPU to the simulation